requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic
streamlit
requests
PyJWT
python-multipart
pdbwhereami
